+++ lib/api/api_rtlmp.dart
@@ -53,8 +53,16 @@ class RtLmp {
       return Response.ok(json.encode(aux), headers: headers);
     });
 
-    /// get daily price for bucket for several ptids
+    /// Get daily price for bucket for several ptids
     /// http://localhost:8080/nyiso/rtlmp/v1/daily/lmp/ptids/61752,61758/start/20170101/end/20170101/bucket/5x16
+    /// Return a list of maps with elements of this form
+    /// ```
+    /// {
+    ///   'date': '2021-01-01',
+    ///   'ptid': 61752,
+    ///   'lmp': 32.3194,
+    /// },
+    /// ```
     router.get(
         '/daily/<component>/ptids/<ptids>/start/<start>/end/<end>/bucket/<bucket>',
         (Request request, String component, String ptids, String start,
@@ -69,20 +77,71 @@ class RtLmp {
       return Response.ok(json.encode(aux), headers: headers);
     });
 
-    /// get daily 7x24 price for all ptids between a start and end date
+
+
+    /// get daily 7x24 price for ALL ptids between a start and end date
     router.get('/daily/mean/<component>/start/<start>/end/<end>',
         (Request request, String component, String start, String end) async {
       var aux = await dailyPriceByPtid(component, start, end);
       return Response.ok(json.encode(aux), headers: headers);
     });
 
-    /// get hourly prices for one ptid
+    /// Get hourly prices for one ptid
+    /// Return a Map in this form
+    /// ```
+    /// {
+    ///   '2021-01-01': <num>[45.67, 32.74, ...],
+    ///   '2021-01-02': <num>[42.83, 37.35, ...],
+    ///   ...
+    /// }
+    /// ```
     router.get('/hourly/<component>/ptid/<ptid>/start/<start>/end/<end>',
         (Request request, String component, String ptid, String start,
             String end) async {
       var aux = await getHourlyPrices(component, int.parse(ptid), start, end);
       return Response.ok(json.encode(aux), headers: headers);
     });
+
+
+    /// Get hourly prices for multiple ptids
+    /// Return a list of maps with elements like
+    /// ```
+    /// {
+    ///   'hourBeginning': '2021-01-01 00:00:00-05:00', '61754': 19.14, '61757': 19.78
+    /// },
+    /// {
+    ///   'hourBeginning': '2021-01-01 01:00:00-05:00', '61754': 20.47, '61757': 20.92
+    /// }
+    /// ```
+    router.get(
+        '/hourly/<component>/ptids/<ptids>/start/<start>/end/<end>',
+            (Request request, String component, String ptids, String start, String end) async {
+          var ptids0 = ptids.split(',').map((e) => int.parse(e)).toList();
+          var startDate = Date.parse(start, location: UTC);
+          var endDate = Date.parse(end, location: UTC);
+
+          var aux = await getHourlyDataSeveral(ptids0, startDate, endDate, component);
+          var out = <Map<String, dynamic>>[];
+
+          var groups = groupBy(aux, (Map e) => e['date']);
+          for (var yyyymmdd in groups.keys) {
+            var date = Date.fromIsoString(yyyymmdd, location: IsoNewEngland.location);
+            var hours = date.hours();
+            var group = groups[yyyymmdd]!;
+            for (var i=0; i<hours.length; i++) {
+              var one = <String,dynamic>{
+                'hourBeginning': hours[i].start.toIso8601String(),
+              };
+              for (var e in group) {
+                one[e['ptid'].toString()] = e[component][i];
+              }
+              out.add(one);
+            }
+          }
+
+          return Response.ok(json.encode(out), headers: headers);
+        });
+
 
     /// Get all the existing ptids in the collection, sorted
     router.get('/ptids', (Request request) async {
@@ -93,6 +152,32 @@ class RtLmp {
 
     return router;
   }
+
+  // Future<void> booboo() async {
+  //   var aux = await getHourlyDataSeveral([61757, 61754], Date.utc(2021, 1, 1),
+  //       Date.utc(2021, 1, 2), 'lmp');
+  //   var out = <Map<String, dynamic>>[];
+  //
+  //   var groups = groupBy(aux, (Map e) => e['date']);
+  //   for (var yyyymmdd in groups.keys) {
+  //     var date = Date.fromIsoString(yyyymmdd, location: IsoNewEngland.location);
+  //     var hours = date.hours();
+  //     var group = groups[yyyymmdd]!;
+  //     for (var i=0; i<hours.length; i++) {
+  //       var one = <String,dynamic>{
+  //         'hourBeginning': hours[i].start.toIso8601String(),
+  //       };
+  //       for (var e in group) {
+  //         one[e['ptid'].toString()] = e['lmp'][i];
+  //       }
+  //       out.add(one);
+  //     }
+  //   }
+  //   print(out);
+  //
+  // }
+
+
 
   /// Return a list with elements:
   /// ```