+++ test/db/nyiso/rt_lmp_hourly_test.dart
@@ -1,0 +1,218 @@
+library test.db.nyiso.rt_lmp_hourly_test;
+
+import 'dart:convert';
+import 'package:elec_server/api/api_rtlmp.dart';
+import 'package:elec_server/src/db/lib_prod_dbs.dart';
+import 'package:elec_server/src/db/nyiso/rt_lmp_hourly.dart';
+import 'package:test/test.dart';
+import 'package:http/http.dart' as http;
+
+import 'package:timezone/data/latest.dart';
+import 'package:date/date.dart';
+import 'package:elec/elec.dart';
+
+Future<void> tests(String rootUrl) async {
+  group('Nyiso RT LMP db tests: ', () {
+    var archive = NyisoRtLmpHourlyArchive();
+    setUp(() async {
+      await archive.dbConfig.db.open();
+    });
+    tearDown(() async {
+      await archive.dbConfig.db.close();
+    });
+    test('process RT hourly lmp report for 2021-01-01 (zones + gen nodes)',
+        () async {
+      var res = archive.processDay(Date.utc(2021, 1, 1));
+      expect(res.length, 580); // both zones and gen nodes
+      expect(res.first.keys.toSet(),
+          {'date', 'ptid', 'lmp', 'congestion', 'losses'});
+      expect(res.first['ptid'], 61757);
+      expect(res.first['date'], '2021-01-01');
+      expect(res.first['lmp'].take(3), [45.67, 32.74, 22.22]);
+    });
+  });
+  group('Nyiso RT LMP api tests: ', () {
+    var api = RtLmp(DbProd.nyiso, iso: Iso.newYork);
+    setUp(() async => await DbProd.nyiso.open());
+    tearDown(() async => await DbProd.nyiso.close());
+    test('get hourly lmp data for one ptid for 2 days', () async {
+      var aux = await api.getHourlyData(
+          61757, Date.utc(2021, 1, 1), Date.utc(2021, 1, 2), 'lmp');
+      expect(aux.length, 2);
+      expect(aux['2021-01-01']!.take(3), [45.67, 32.74, 22.22]);
+      var url = '$rootUrl/nyiso/rtlmp/v1/hourly/lmp/'
+          'ptid/61757/start/2021-01-01/end/2021-01-02';
+      var res = await http
+          .get(Uri.parse(url), headers: {'Content-Type': 'application/json'});
+      var data = json.decode(res.body) as Map;
+      expect(data.length, 2);
+      expect(data['2021-01-01']!.take(3), [45.67, 32.74, 22.22]);
+    });
+
+
+    test('get hourly lmp data for several ptids for several days', () async {
+      var url = '$rootUrl/nyiso/rtlmp/v1/hourly/lmp/'
+          'ptids/61757,61754/start/2021-01-01/end/2021-01-02';
+      var res = await http
+          .get(Uri.parse(url), headers: {'Content-Type': 'application/json'});
+      var data = json.decode(res.body) as List;
+      expect(data.length, 48);
+      expect(data.first, {
+        'hourBeginning': '2021-01-01T00:00:00.000-0500',
+        '61754': 8.49,
+        '61757': 45.67,
+      });
+    });
+
+
+    test('get daily lmp prices by peak bucket for one ptid', () async {
+      var res = await http.get(
+          Uri.parse('$rootUrl/nyiso/rtlmp/v1/daily/lmp/'
+              'ptid/61757/start/2021-01-01/end/2021-01-07/bucket/2x16H'),
+          headers: {'Content-Type': 'application/json'});
+      var aux = json.decode(res.body) as List;
+      expect(aux.length, 3);
+      expect(aux.first, {'date': '2021-01-01', 'lmp': 45.218125});
+    });
+
+
+    test('get daily lmp prices by peak bucket for several ptids', () async {
+      var data = await api.getDailyBucketPriceSeveral('lmp', [61757, 61752],
+          Date.utc(2021, 1, 1), Date.utc(2021, 1, 7), Bucket.b2x16H);
+      expect(data.length, 6);
+      var n57 = data
+          .firstWhere((e) => e['ptid'] == 61757 && e['date'] == '2021-01-01');
+      expect(n57,
+          {'ptid': 61757, 'date': '2021-01-01', 'lmp': 45.218125});
+      var res = await http.get(
+          Uri.parse('$rootUrl/nyiso/rtlmp/v1/daily/lmp/'
+              'ptids/61757,61752/start/2021-01-01/end/2021-01-07/bucket/2x16H'),
+          headers: {'Content-Type': 'application/json'});
+      var aux = json.decode(res.body) as List;
+      expect(aux.length, 6);
+      expect(
+          aux.firstWhere(
+              (e) => e['ptid'] == 61757 && e['date'] == '2021-01-01'),
+          {'ptid': 61757, 'date': '2021-01-01', 'lmp': 45.218125});
+    });
+
+    test('get daily lmp prices by flat bucket', () async {
+      var data = await api.getDailyBucketPriceSeveral('lmp', [61757],
+          Date.utc(2021, 1, 1), Date.utc(2021, 1, 7), Bucket.atc);
+      expect(data[0]['lmp'], 41.035833333333336);
+      expect(data.length, 7);
+    });
+
+    test('get daily 7x24 prices direct method (calculated by mongo)', () async {
+      var data = await api.getDailyAtcPrices([61757, 61758],
+          Date.utc(2021, 1, 1), Date.utc(2021, 1, 31), 'congestion');
+      expect(data.first, {
+        'date': '2021-01-01',
+        'ptid': 61757,
+        'congestion': -18.047916666666666,
+      });
+      expect(data.length, 62);
+      var url = '$rootUrl/nyiso/rtlmp/v1/daily/congestion/ptids/61752,61758/start/2021-01-01/end/2021-01-31/bucket/7x24';
+      var aux = await http.get(Uri.parse(url));
+      var res = json.decode(aux.body) as List;
+      expect(res.length, 62);
+    });
+
+    test('get monthly 7x24 prices direct method (calculated by mongo)',
+        () async {
+      var data = await api.getMonthlyAtcPrices([61752, 61758],
+          Month.utc(2021, 1), Month.utc(2021, 12), 'congestion');
+      expect(data.first, {
+        'month': '2021-01',
+        'ptid': 61752,
+        'congestion': -2.4525,
+      });
+      expect(data.length, 24);
+    });
+
+    test('get mean daily 7x24 prices all nodes 2021-01-01 (mongo)', () async {
+      var data = await api.dailyPriceByPtid('lmp', '2021-01-01', '2021-01-01');
+      var capitl = data.firstWhere((e) => e['ptid'] == 61757);
+      expect((capitl['lmp'] as num).toStringAsFixed(4), '41.0358');
+      expect(data.length, 580);
+    });
+    test('get monthly lmp prices by flat bucket', () async {
+      var data = await api.getMonthlyBucketPrice(
+          'lmp', 61757, '202101', '202102', 'flat');
+      expect(data.length, 2);
+      expect(data.first, {'month': '2021-01', 'lmp': 41.377916666666664});
+    });
+  });
+  // group('Nyiso DAM LMP speed tests: ', () {
+  //   var api = DaLmp(DbProd.nyiso, iso: Iso.newYork);
+  //   var sw = Stopwatch();
+  //   setUp(() async => await DbProd.nyiso.open());
+  //   tearDown(() async => await DbProd.nyiso.close());
+  //   test('hourly lmp', () async {
+  //     /// 43 ms for one ptid 365 days, 55 ms for 700 days.
+  //     /// 28 ms just the db query to select one node for 365 days, 40 ms for 700 days.
+  //     sw.start();
+  //     var aux = await api.getHourlyData(
+  //         61757, Date.utc(2019, 2, 14), Date.utc(2020, 1, 13), 'congestion');
+  //     sw.stop();
+  //     expect(sw.elapsedMilliseconds, lessThan(100));
+  //     expect(aux.containsKey('2019-02-14'), true);
+  //   });
+  //   test('get mean daily lmp prices all nodes 2017-01-01 (mongo)', () async {
+  //     sw.reset();
+  //     sw.start();
+  //     var data = await api.dailyPriceByPtid('lmp', '2019-02-14', '2020-02-13');
+  //     sw.stop();
+  //     expect(sw.elapsedMilliseconds, lessThan(3000));
+  //     expect(data.length, 208891);
+  //   });
+  // });
+
+  // group('Nyiso RT LMP client tests: ', () {
+  //   var daLmp = client.DaLmp(http.Client(), iso: Iso.newYork, rootUrl: rootUrl);
+  //   test('get daily peak price between two dates', () async {
+  //     var data = await daLmp.getDailyLmpBucket(61752, LmpComponent.lmp,
+  //         Bucket.b5x16, Date.utc(2019, 1, 1), Date.utc(2019, 1, 5));
+  //     expect(data.length, 3);
+  //     expect(data.toList(), [
+  //       IntervalTuple(Date(2019, 1, 2, location: location), 31.678124999999998),
+  //       IntervalTuple(Date(2019, 1, 3, location: location), 29.316875000000003),
+  //       IntervalTuple(Date(2019, 1, 4, location: location), 23.630625000000002),
+  //     ]);
+  //   });
+  //   test('get monthly peak price between two dates', () async {
+  //     var data = await daLmp.getMonthlyLmpBucket(61752, LmpComponent.congestion,
+  //         IsoNewEngland.bucket5x16, Month.utc(2019, 1), Month.utc(2019, 12));
+  //     expect(data.length, 12);
+  //     expect(
+  //         data.first,
+  //         IntervalTuple(
+  //             Month(2019, 1, location: location), -7.513068181818175));
+  //   });
+  //   test('get hourly price for 2017-01-01', () async {
+  //     var data = await daLmp.getHourlyLmp(
+  //         61752, LmpComponent.lmp, Date.utc(2019, 1, 1), Date.utc(2019, 1, 1));
+  //     expect(data.length, 24);
+  //     expect(
+  //         data.first,
+  //         IntervalTuple(
+  //             Hour.beginning(TZDateTime(location, 2019, 1, 1)), 11.83));
+  //   });
+  //   test('get daily prices all nodes', () async {
+  //     var data = await daLmp.getDailyPricesAllNodes(
+  //         LmpComponent.lmp, Date.utc(2019, 1, 1), Date.utc(2019, 1, 3));
+  //     expect(data.length, 570);
+  //     var zA = data[61752]!;
+  //     expect(zA.first.value, 17.090833333333332);
+  //   });
+  // });
+}
+
+
+void main() async {
+  initializeTimeZones();
+
+  DbProd();
+  // dotenv.load('.env/prod.env');
+  tests('http://127.0.0.1:8080');
+}